# docs/ai-tools-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DashAdmin AI Tools API",
    "version": "2.0.0",
    "description": "API for AI agents (n8n) to query DashAdmin data. Requires bot authentication via headers."
  },
  "servers": [
    {
      "url": "https://www.mydashadmin.ru",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "BotAuth": []
    }
  ],
...
  "paths": {
    "/api/ai-tools/select-club": {
      "get": {
        "summary": "Get Available Clubs",
        "description": "Returns the list of clubs available to the authenticated user and their current selection.",
        "operationId": "getAvailableClubs",
        "responses": {
          "200": {
            "description": "Club information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClubSelection" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Select Club",
        "description": "Sets the active club for subsequent AI queries.",
        "operationId": "selectClub",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SelectClubRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Club selected successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClubSelection" }
              }
            }
          }
        }
      }
    },
    "/api/ai-tools/revenue": {
      "get": {
        "summary": "Get Revenue Report",
        "description": "Fetches revenue data for the selected club. Supports various time periods.",
        "operationId": "getRevenue",
        "parameters": [
          {
            "name": "clubId",
            "in": "query",
            "required": false,
            "description": "Override club ID (defaults to selected club)",
            "schema": { "type": "string" }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "description": "Predefined time period",
            "schema": {
              "type": "string",
              "enum": ["today", "yesterday", "last7days", "last30days", "this_month"]
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "description": "Custom start date (YYYY-MM-DD)",
            "schema": { "type": "string" }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "description": "Custom end date (YYYY-MM-DD)",
            "schema": { "type": "string" }
          },
          {
            "name": "shiftType",
            "in": "query",
            "required": false,
            "description": "Filter by shift type",
            "schema": { "type": "string", "enum": ["DAY", "NIGHT"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RevenueReport" }
              }
            }
          }
        }
      }
    },
    "/api/ai-tools/shifts-summary": {
      "get": {
        "summary": "Get Shifts Summary",
        "description": "Fetches summary statistics for shifts including hours, counts, and employee data.",
        "operationId": "getShiftsSummary",
        "parameters": [
          {
            "name": "clubId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["today", "yesterday", "last7days", "last30days", "this_month"]
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter by shift status",
            "schema": { "type": "string", "enum": ["ACTIVE", "CLOSED"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Shifts summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShiftsSummary" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BotAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Messenger-User-Id",
        "description": "Bot authentication via linked user account"
      }
    },
    "schemas": {
      "ClubSelection": {
        "type": "object",
        "properties": {
          "selected_club": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": { "type": "integer" },
              "name": { "type": "string" }
            }
          },
          "available_clubs": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "name": { "type": "string" }
              }
            }
          }
        }
      },
      "SelectClubRequest": {
        "type": "object",
        "required": ["clubId"],
        "properties": {
          "clubId": { "type": "string", "description": "ID of club to select" }
        }
      },
      "RevenueReport": {
        "type": "object",
        "properties": {
          "club_id": { "type": "integer" },
          "club_name": { "type": "string" },
          "period": {
            "type": "object",
            "properties": {
              "start": { "type": "string" },
              "end": { "type": "string" },
              "label": { "type": "string" }
            }
          },
          "revenue": {
            "type": "object",
            "properties": {
              "total": { "type": "number" },
              "expenses": { "type": "number" },
              "net": { "type": "number" },
              "cash": { "type": "number" },
              "card": { "type": "number" }
            }
          },
          "shifts": {
            "type": "object",
            "properties": {
              "count": { "type": "integer" },
              "total_hours": { "type": "number" },
              "employees": { "type": "integer" }
            }
          }
        }
      },
      "ShiftsSummary": {
        "type": "object",
        "properties": {
          "club_id": { "type": "integer" },
          "club_name": { "type": "string" },
          "period": { "$ref": "#/components/schemas/RevenueReport/properties/period" },
          "summary": {
            "type": "object",
            "properties": {
              "total_shifts": { "type": "integer" },
              "total_hours": { "type": "number" },
              "average_duration_hours": { "type": "number" },
              "day_shifts": { "type": "integer" },
              "night_shifts": { "type": "integer" },
              "active_shifts": { "type": "integer" },
              "closed_shifts": { "type": "integer" },
              "unique_employees": { "type": "integer" }
            }
          }
        }
      },
...
    }
  }
}